refactor(front): Reorganize component structure - Move UserContext to Common module - Move UpperInfoBox out of Common module
